Ensure that exceptions are always retrieved from tasks
In particular wrap ensure_future to ensure that there's always a done
callback that will (at least) propagate unhandled exceptions to the
event loop.

This will result in crashes. Obviously this could be changed to log
errors and move on, but this should always be a bug...so crashing seems
like a sensible default.